==> 6_estimate_biomarker_associations.Rmd <==
---
title: "Estimate biomarker associations"
author: "Matt Ploenzke"
output: html_document
---

Packages

```{r}
library(readxl)
library(tidyverse)
library(PharmacoGx)
```

Set options 

```{r}
comparison <- 'CCLE_GDSC' # CLE_GDSC1000_CTRPv2_FIMM_gCSI
pset_dir <- file.path('~/Downloads', 'PSets')
posteriors_dir <- file.path('~/Downloads','results')
use_all_features <- TRUE # use all available features (TRUE) or restrict to features with a previously-documented association (FALSE)
downsample_features <- TRUE # downsample to 5K most variable features (TRUE, recommended)
feature.type <- "rna" # mutation fusion cnv rna rna2 rnaseq
impute_with_mean <- TRUE # if a feature exists for some cells but not others, replace the missings with the mean
mean_common_cells <- TRUE # reduce joint dataset to mean sensitivity measure 
include_individual_posteriors <- TRUE # for individual study comparisons, should the posterior fit via only that dataset be included (TRUE) or should it be compared to the joint fit posterior (FALSE)
individual_study_measures <- c('waterfall','naive','auc_recomputed','posterior_probability_sensitive','cell_type')
joint_study_measures <- c('naive','auc_recomputed','posterior_probability_sensitive','cell_type')
output_dir <- file.path(comparison,'biomarkers')
dir.create(output_dir,showWarnings = FALSE, recursive=TRUE)
```

Load posteriors

```{r}
if (include_individual_posteriors) {
  individual_posteriors <- readRDS(file.path(posteriors_dir,comparison,'fit_within_study', 'posterior.formatted.RDS'))
}
joint_posteriors <- readRDS(file.path(posteriors_dir,comparison,'fit_across_study', 'posterior.formatted.RDS')) 
gene_drug_asociations <- read_excel(file.path(pset_dir,'gene_drug_asociations.xlsx'), sheet = "mutation") %>%
  as_tibble() %>%
  bind_rows(read_excel(file.path(pset_dir,'gene_drug_asociations.xlsx'), sheet = "copy_number_variation"),
            read_excel(file.path(pset_dir,'gene_drug_asociations.xlsx'), sheet = "expression"))
if (include_individual_posteriors) {
  all_posteriors <- individual_posteriors %>% 
    mutate(Model_fit = 'individual') %>%
    bind_rows(joint_posteriors %>%
                mutate(Model_fit = 'joint'))
  rm(individual_posteriors, joint_posteriors)
} else {
  all_posteriors <- joint_posteriors %>% mutate(Model_fit = 'joint')
  rm(joint_posteriors)
}
gene_drug_asociations <- gene_drug_asociations %>%
  mutate(type=case_when(type=='expression' ~ 'rna',
                        type=='copy_number_variation' ~ 'cnv',
                        TRUE ~ type)) %>%
  rename(drug=compound) %>%
  select(type, drug, gene, source)
gene_drug_asociations <- gene_drug_asociations %>% 
  bind_rows(gene_drug_asociations %>% 
              filter(type=='rna') %>%
              mutate(type='rna2')) %>%
  bind_rows(gene_drug_asociations %>% 
              filter(type=='rna') %>%
              mutate(type='rnaseq'))
```

Pull list of drugs to assess

```{r}
all_drugs <- all_posteriors %>% 
  group_by(drug) %>% 
  distinct(experiment) %>% 
  filter(n()>1) %>% 
  ungroup() %>% 
  distinct(drug) %>% 
  pull(drug) 
broad_drugs <- c('17-AAG','paclitaxel','PD-0325901','AZD6244')
posterior <- all_posteriors %>% select(cell, drug, experiment, cell_type, posterior_probability_sensitive, Model_fit)
datasets <- posterior %>% distinct(experiment) %>% pull()
gene_drug_asociations <- gene_drug_asociations %>% filter(drug %in% all_drugs)
```

Load gene names

```{r}
ensembl <- try(biomaRt::useEnsembl(biomart="ensembl", dataset="hsapiens_gene_ensembl"), silent=TRUE)
if (class(ensembl) == 'try-error') {
  ensembl = biomaRt::useEnsembl(biomart="ensembl", dataset="hsapiens_gene_ensembl",mirror='uswest')
}
chr_genes <- biomaRt::getBM(attributes=c('ensembl_gene_id','ensembl_transcript_id',
                                         'hgnc_symbol','entrezgene_id'), mart = ensembl) %>%
  as_tibble() %>% 
  distinct(ensembl_gene_id, hgnc_symbol,entrezgene_id)
known_ensembl_ids <- gene_drug_asociations %>% 
  distinct(gene) %>% 
  left_join(chr_genes, by=c('gene'='hgnc_symbol')) %>% 
  filter(!is.na(ensembl_gene_id)) %>% 
  pull(ensembl_gene_id)
```

Set up X and Y data matrices, one per study

```{r}
rez <- list()
for (dataset in datasets) {
  print(dataset)
  load(file.path(pset_dir,paste(dataset,'RData',sep='.')))
  eval(parse(text = paste('tempPset <- ',dataset,sep='')))
  print(paste(dataset, 'available features:',paste(mDataNames(tempPset),collapse=' '), sep=' '))
  y_new_temp <- summarizeSensitivityProfiles(tempPset, sensitivity.measure = 'auc_recomputed', summary.stat = 'median', verbose = FALSE) %>%
    as.data.frame() %>%
    rownames_to_column('drug') %>%
    as_tibble() %>% 
    gather(key='cell',value='value',-drug) %>%
    group_by(drug) %>%
    filter(!is.na(value)) %>%
    mutate(waterfall=PharmacoGx:::callingWaterfall(value,type='AUC')) %>%
    ungroup() %>% 
    rename(auc_recomputed=value) %>%
    mutate(experiment = dataset)
  ii <- 1
  any.features.exist <- FALSE
  for (feat in feature.type) {
    if (!(feat %in% mDataNames(tempPset))) {
      next()  
    }
    print(feat)
    features <- gene_drug_asociations %>%
      filter(type==feat) %>%
      distinct(gene) %>% 
      pull()
    if (use_all_features) {
      if ('Symbol' %in% colnames(featureInfo(tempPset,feat))) {
        idx <- 1:length(featureInfo(tempPset,feat)$Symbol)
      } else if ('gene_name' %in% colnames(featureInfo(tempPset,feat))) {
        idx <- 1:length(featureInfo(tempPset,feat)$gene_name)
      } else {
        if (nrow(featureInfo(tempPset,feat))>0) {
          idx <- 1:nrow(featureInfo(tempPset,feat))
        } else {
          print(paste('Unknown column in PSet: ', dataset, sep=''))
          next()
        }
      }
      any.features.exist <- TRUE
    } else {
      if ('Symbol' %in% colnames(featureInfo(tempPset,feat))) {
        idx <- which(featureInfo(tempPset,feat)$Symbol %in% features)
        any.features.exist <- TRUE
      } else if ('gene_name' %in% colnames(featureInfo(tempPset,feat))) {
        idx <- which(featureInfo(tempPset,feat)$gene_name %in% features)
        any.features.exist <- TRUE
      } else {
        print(paste('Unknown column in PSet: ', dataset, sep=''))
        next()
      }
    }
    molfeatures <- try(fNames(tempPset, feat)[idx],silent=TRUE)
    if (class(molfeatures)=='try-error') {
      x_new <- tibble(cell='NA')
    } else {
      molfeatures <- molfeatures[!is.na(molfeatures)]
      if (length(molfeatures) > 0) {
        stat <- ifelse(feat %in% c('mutation','fusion'),'and','median')
        x_new <- summarizeMolecularProfiles(tempPset, mDataType = feat, summary.stat = stat, verbose=TRUE) %>%
          Biobase::exprs() %>%
          as.data.frame() %>%
          rownames_to_column('gene') %>%
          as_tibble() %>%
          mutate_at(vars(-gene),as.character) %>%
          gather(cell, value, -gene) %>%
          filter(gene %in% molfeatures) %>%
          mutate(gene = paste(gene,feat,sep='_')) %>%
          spread(gene, value) 
        if (impute_with_mean) {
          colsumss <- x_new %>% select(-cell) %>% mutate_all(as.numeric) %>% summarise_all(sum, na.rm =TRUE)
          for_removal <- names(colsumss)[colsumss == 0]
          x_new <- x_new %>% 
            select(-for_removal) %>%
            mutate_at(vars(-cell), as.numeric) %>%
            mutate_at(vars(-cell), function(ii) {
              replace(ii, is.na(ii),mean(ii,na.rm=TRUE))
            })
        } else {
          x_new <- x_new %>% 
          na.omit() 
        }
      } else {
        x_new <- tibble(cell='NA')
      }
    }
    if (ii == 1) {
      joined <- x_new 
      ii <- ii+1
    } else {
      joined <- x_new %>% full_join(joined,by='cell')
    }
    if (downsample_features) {
      var.sds <- c()
      for (feat in feature.type) { 
        var.sds <- c(var.sds, apply(joined %>% select(contains(feat)), 2, sd, na.rm=TRUE))
      }
      var.sds <- sort(var.sds,decreasing = TRUE)
      top.variable <- names(var.sds[1:min(length(var.sds),5000)])
      top.variable <- unique(c(top.variable, paste0(known_ensembl_ids, '_at_rna')))
      top.variable <- top.variable[top.variable %in% colnames(joined)]
      joined <- joined %>% select(cell, top.variable)
    }
  }
  if (any.features.exist) {
    rez[[dataset]] <- joined %>% inner_join(y_new_temp, by='cell') 
  } else {
    rez[[dataset]] <- y_new_temp
  }
  if (include_individual_posteriors) {
    rez[[dataset]] <- rez[[dataset]] %>%
      inner_join(posterior %>% 
               filter(Model_fit == 'individual') %>% 
               select(-Model_fit) %>%
               distinct(),by=c('cell','experiment','drug')) %>% 
      mutate(cell_type = case_when(posterior_probability_sensitive > .5 ~ 'sensitive',
                               TRUE ~ 'resistant')) 
  } else {
    rez[[dataset]] <- rez[[dataset]] %>%
      inner_join(posterior %>% 
             select(-Model_fit) %>%
             distinct(),by=c('cell','experiment','drug')) %>% 
      mutate(cell_type = case_when(posterior_probability_sensitive > .5 ~ 'sensitive',
                             TRUE ~ 'resistant')) 
  }
  eval(parse(text=paste('rm(',dataset,',tempPset, x_new, joined, y_new_temp)',sep='')))
}
```

Fit models for varying outcome variables

```{r}
all_effect_sizes <- tibble()
for (drug_for_biomarkers in all_drugs) {
  print(drug_for_biomarkers)
  naive_cutoff <- ifelse(drug_for_biomarkers %in% broad_drugs,.4,.2)
  for (study in names(rez)) {
    rez[[study]] <- rez[[study]][!(colSums(is.na(rez[[study]])) == nrow(rez[[study]]))]
    ourData <- rez[[study]] %>%
      na.omit() %>%
      filter(drug==drug_for_biomarkers) %>%
      mutate(naive = ifelse(auc_recomputed>naive_cutoff,'sensitive','resistant'))
    if (nrow(ourData)==0) {next()}
    for (depVar in individual_study_measures) {
      if (depVar == 'waterfall') { 
        ourData <- ourData %>% filter(waterfall!='intermediate')
      }
      
      # gather expression data
      expr_data <- ourData %>% 
        select(-drug, -cell, -experiment, #-waterfall, 
               -auc_recomputed, -naive, -cell_type, 
               -posterior_probability_sensitive) %>%
        mutate_all(as.numeric) %>%
        t() 
      if (nrow(expr_data) == 0) {next()}
      expr_data <- expr_data[!duplicated(rownames(expr_data)),]
      colnames(expr_data) <- pull(ourData %>% select(cell))
      
      # featureData
      fdata <- tibble(Feature = rownames(expr_data)) %>%
        mutate(Full_Feature = Feature) %>%
        separate(Feature,into=c('Feature','extra1','extra2'), fill='right',sep='_') %>% 
        left_join(chr_genes,by = c('Feature'='ensembl_gene_id')) %>%
        group_by(Full_Feature) %>%
        filter(row_number()==1) %>%
        ungroup() %>%
        mutate(symbol = ifelse(is.na(hgnc_symbol),Feature, hgnc_symbol)) %>%
        mutate(type = ifelse(is.na(extra2),extra1, extra2)) %>%
        select(Full_Feature, symbol, Feature, type, entrezgene_id) %>%
        distinct(Full_Feature, .keep_all=TRUE) %>%
        column_to_rownames('Full_Feature') %>%
        Biobase::AnnotatedDataFrame()
      
      # pull pheno data
      tmp <- ourData %>% 
        select(cell, depVar) %>% 
        rename(Sensitivity= depVar) %>%
        distinct(Cell, .keep_all=TRUE) %>%
        column_to_rownames('cell')
      if (depVar %in% c('waterfall','naive','cell_type')) {
        tmp$Sensitivity <- as.factor(as.character(tmp$Sensitivity))
      }
      pdata <- Biobase::AnnotatedDataFrame(tmp)
      
      # create expression set
      eset <- new("ExpressionSet", exprs = as.matrix(expr_data), phenoData = pdata, featureData=fdata)
      
      # fit model
      my_design <- try(model.matrix(~ eset$Sensitivity),silent=TRUE)
      if (class(my_design) == 'try-error') {browser()}
      if (class(my_design) != 'try-error') {
        fit <- try(limma::lmFit(object=eset, design=my_design),silent=TRUE)
        if (class(fit) == 'try-error') {browser()}
        if (class(fit) != 'try-error') {
          fit <- limma::eBayes(fit)
      
          all_effect_sizes <- as_tibble(limma::topTable(fit, number = Inf)) %>% 
              mutate(Measure = depVar,
                     Dataset = study, 
                     Drug = drug_for_biomarkers) %>% 
              bind_rows(all_effect_sizes)
        }
      }
    }
  }
}
```

Same thing for the joint posterior

```{r}
for (drug_for_biomarkers in all_drugs) {
  print(drug_for_biomarkers)
  naive_cutoff <- ifelse(drug_for_biomarkers %in% broad_drugs,.4,.2)
  if (mean_common_cells) {
    experi_data <- posterior %>% 
      distinct(cell, experiment) %>% 
      mutate(experiment2=experiment) %>%
      spread(experiment,experiment2) %>%
      replace(is.na(.), '') %>% 
      group_by(cell) %>%
      mutate_at(vars(-group_cols()), funs(paste(.,'_', sep=''))) %>%
      ungroup() %>%
      mutate_all(funs(case_when(.=='_' ~ '', TRUE~.))) %>%
      unite(experiment, CCLE, CTRPv2, FIMM, gCSI, GDSC1000, sep='') %>%
      mutate(experiment = substr(experiment, 1, nchar(experiment)-1))
    ourData <- do.call(bind_rows, rez) %>% 
      ungroup() %>%
      select(-experiment, -cell_type, -posterior_probability_sensitive) %>%
      filter(drug==drug_for_biomarkers) %>%
      group_by(cell, drug) %>%
      mutate_at(vars(-group_cols()), as.numeric) %>%
      summarise_all(mean, na.rm = TRUE) %>%
      ungroup() %>%
      na.omit() %>%
      inner_join(posterior %>% 
                   filter(Model_fit == 'joint',drug==drug_for_biomarkers) %>%
                   select(-Model_fit, -experiment) %>%
                   distinct(), by=c('cell','drug')) %>%
      distinct() %>%
      mutate(naive = ifelse(auc_recomputed>naive_cutoff,'sensitive','resistant')) %>%
      mutate(cell_type = case_when(posterior_probability_sensitive > .5 ~ 'sensitive',
                                 TRUE ~ 'resistant')) %>%
      group_by(cell) %>%
      mutate(rownum = row_number()) %>%
      mutate(rownum = ifelse(rownum>1,paste('_',rownum,sep=''),'')) %>%
      ungroup() %>%
      unite(cell, cell,rownum,sep='') %>%
      left_join(experi_data, by='cell')
  } else {
    ourData <- do.call(bind_rows, rez) %>% 
      ungroup() %>%
      select(-experiment, -cell_type, -posterior_probability_sensitive, -auc_recomputed) %>%
      filter(drug==drug_for_biomarkers) %>%
      group_by(cell, drug) %>%
      mutate_at(vars(-group_cols()), as.numeric) %>%
      summarise_all(mean, na.rm = TRUE) %>%
      ungroup() %>%
      right_join(do.call(bind_rows, rez) %>% 
                   select(cell, drug, experiment, auc_recomputed) %>%
                   filter(drug==drug_for_biomarkers), by=c('cell','drug')) %>%
      na.omit() %>%
      inner_join(posterior %>% 
                   filter(Model_fit == 'joint',drug==drug_for_biomarkers) %>%
                   select(-Model_fit) %>%
                   distinct(), by=c('cell','drug', 'experiment')) %>%
      distinct() %>%
      mutate(naive = ifelse(auc_recomputed>naive_cutoff,'sensitive','resistant')) %>%
      mutate(cell_type = case_when(posterior_probability_sensitive > .5 ~ 'sensitive',
                                 TRUE ~ 'resistant')) %>%
      group_by(cell) %>%
      mutate(rownum = row_number()) %>%
      mutate(rownum = ifelse(rownum>1,paste('_',rownum,sep=''),'')) %>%
      ungroup() %>%
      unite(cell, cell,rownum,sep='') 
  }
  for (depVar in joint_study_measures) {
    # gather expression data
    expr_data <- ourData %>% 
      select(-drug, -cell, -experiment,
             -auc_recomputed, -naive, -cell_type, 
             -posterior_probability_sensitive) %>% 
      mutate_all(as.numeric) %>%
      t() 
    expr_data <- expr_data[!duplicated(rownames(expr_data)),]
    colnames(expr_data) <- pull(ourData %>% select(cell))
    
    # featureData:
    fdata <- tibble(Feature = rownames(expr_data)) %>%
      mutate(Full_Feature = Feature) %>%
      separate(Feature,into=c('Feature','extra1','extra2'), fill='right',sep='_') %>% 
      left_join(chr_genes,by = c('Feature'='ensembl_gene_id')) %>%
      group_by(Full_Feature) %>%
      filter(row_number()==1) %>%
      ungroup() %>%
      mutate(symbol = ifelse(is.na(hgnc_symbol),Feature, hgnc_symbol)) %>%
      mutate(type = ifelse(is.na(extra2),extra1, extra2)) %>%
      select(Full_Feature, symbol, Feature, type,entrezgene_id) %>%
      distinct(Full_Feature, .keep_all=TRUE) %>%
      column_to_rownames('Full_Feature') %>%
      Biobase::AnnotatedDataFrame()
    
    # pull pheno data
    if (length(unique(ourData$cell_type))<2 & depVar=='cell_type') {
      tthresh <- quantile(ourData$posterior_probability_sensitive,probs=.8)
      print(paste0('Only one outcome category defined for ',
                   drug_for_biomarkers,
                   '. Lowering the threshold of sensitive to: ',
                   tthresh))
      ourData <- ourData %>% 
        mutate(cell_type = case_when(posterior_probability_sensitive>=tthresh ~ 'sensitive',
                                     TRUE ~ 'resistant'))
    }
    tmp <- ourData %>% 
      select(cell, depVar) %>% 
      rename(Sensitivity= depVar) %>%
      distinct(cell, .keep_all=TRUE) %>%
      column_to_rownames('cell')
    if (depVar %in% c('naive','cell_type')) {
      tmp$Sensitivity <- as.factor(as.character(tmp$Sensitivity))
    }
    pdata <- Biobase::AnnotatedDataFrame(tmp)
    
    # create expression set
    eset <- new("ExpressionSet", exprs = as.matrix(expr_data), phenoData = pdata, featureData=fdata)
    
    # fit model 
    my_design <- try(model.matrix(~ eset$Sensitivity),silent=TRUE)
    if (class(my_design) != 'try-error') {
      fit <- try(limma::lmFit(object=eset, design=my_design),silent=TRUE)
      if (class(fit) != 'try-error') {
        fit <- limma::eBayes(fit)
        
        all_effect_sizes <- as_tibble(limma::topTable(fit, number = Inf)) %>% 
            mutate(Measure = depVar,
                   Dataset = 'Joint', 
                   Drug = drug_for_biomarkers) %>% 
            bind_rows(all_effect_sizes)
       }
    }
  }
}
```

Format data frame and export

```{r}
all_effect_sizes <- all_effect_sizes %>%
  mutate(Method = case_when(Measure %in% c('cell_type','naive', 'waterfall')  ~ 'Binarized',
                            TRUE ~ 'Continuous')) %>%
  mutate(Measure = case_when(Measure %in% c('auc_recomputed','naive') ~ 'AAC',
                             Measure %in% c('cell_type','posterior_probability_sensitive') ~ 'Posterior',
                             TRUE ~ Measure)) %>% 
  full_join(gene_drug_asociations %>% 
              filter(drug %in% all_drugs) %>%
              rename(lit_type = type)  %>%
              mutate(isKnown = TRUE), by=c('symbol'='gene','Drug'='drug')) %>%
  replace_na(list(isKnown = FALSE,
                  logFC=0,
                  P.val=1))
write_csv(all_effect_sizes, path=file.path(output_dir, paste0(feature.type,'.csv')), 
              col_names = TRUE, append = FALSE)
```
