==> 9_make_supplementary_figures.Rmd <==
---
title: "Make supplementary figures"
author: "Matt Ploenzke"
output: html_document
---

# Set up 

Libraries

```{r}
library(tidyverse)
library(viridis)
library(cowplot)
library(ggrepel)
theme_set(theme_bw())
```

Options

```{r}
results_dir <- '~/Downloads/results'
output_dir <- '~/Downloads/supplemental_figs/'
```

# Dose response curves and IC50 v AAC plots
  
  - These are created in the 2_plot_dose_curves.Rmd
  
# Pearson v Spearman (CCLE-GDSC-CTRP-FIMM-gCSI)

```{r}
intersection.fitted.list <- readRDS(file.path(results_dir,'CCLE_GDSC1000_gCSI_FIMM_CTRPv2','fit_within_study','intersection.fitted.list.RDS'))
twoway_agreement <- tibble()
for (comp in names(intersection.fitted.list)) {
  tibb <- intersection.fitted.list[[comp]]
  curr_studies <- strsplit(comp,'_')[[1]]
  pearson <- tibb %>% 
    select(drug, cell, realized_value, experiment) %>%
    spread(experiment, realized_value) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(pearson = cor(v1,v2,method='pearson'),
              spearman = cor(v1,v2,method='spearman')) %>%
    ungroup()
  twoway_agreement <- pearson %>% 
    mutate(comparison = comp) %>% 
    bind_rows(twoway_agreement)
}
twoway_agreement %>%
  mutate(comparison = gsub('_',' & ', comparison)) %>% 
  ggplot(aes(x=spearman, y=pearson)) +
  geom_abline(aes(slope=1,intercept=0),lty=2, color='grey') +
  geom_point(shape=4) + 
  theme(panel.grid = element_blank(),
        axis.text = element_text(size=14),
        axis.title = element_text(size=14),
        plot.title = element_text(size=16),
        legend.position='none') +
  scale_fill_viridis(discrete = TRUE,direction = -1) +
  scale_x_continuous(breaks=c(-.5,0,.5,1), limits = c(-.5,1)) + 
  scale_y_continuous(breaks=c(-.5,0,.5,1), limits = c(-.5,1)) + 
  labs(x='Spearman rank correlation', y='Pearson correlation', title='Common drugs and cells correlation') +
  facet_wrap(vars(comparison),ncol=5)
ggsave(file.path(output_dir, 'pearson_spearman_all_datasets.pdf'),width=14,height=7,units='in')
```

# Pearson v Spearman (PRISM-CTRP-GDSC)

```{r}
intersection.fitted.list <- readRDS(file.path(results_dir,'PRISM_GDSC_CTRP','zeros_included','intersection.fitted.list.RDS')) 
twoway_agreement <- tibble()
for (comp in names(intersection.fitted.list)) {
  tibb <- intersection.fitted.list[[comp]]
  curr_studies <- strsplit(comp,'_')[[1]]
  pearson <- tibb %>% 
    filter(drug_type=='broad') %>%
    select(drug, cell, realized_value, experiment) %>%
    spread(experiment, realized_value) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(pearson = cor(v1,v2,method='pearson'),
              spearman = cor(v1,v2,method='spearman')) %>%
    ungroup()
  twoway_agreement <- pearson %>% 
    mutate(comparison = comp) %>% 
    bind_rows(twoway_agreement)
}
twoway_agreement %>%
  filter(comparison != 'CTRP_GDSC') %>% 
  mutate(comparison = gsub('_',' & ', comparison)) %>% 
  ggplot(aes(x=spearman, y=pearson)) +
  geom_abline(aes(slope=1,intercept=0),lty=2, color='grey') +
  geom_point(shape=4) + 
  theme(panel.grid = element_blank(),
        axis.text = element_text(size=14),
        axis.title = element_text(size=14),
        plot.title = element_text(size=16),
        legend.position='none') +
  scale_fill_viridis(discrete = TRUE,direction = -1) +
  scale_x_continuous(breaks=c(-.5,0,.5,1), limits = c(-.5,1)) + 
  scale_y_continuous(breaks=c(-.5,0,.5,1), limits = c(-.5,1)) + 
  labs(x='Spearman rank correlation', y='Pearson correlation',title='Common drugs and cells correlation - PRISM-CTRP-GDSC') +
  facet_wrap(vars(comparison),ncol=5)
ggsave(file.path(output_dir, 'pearson_spearman_prism_datasets.pdf'),width=14,height=7,units='in')
```

# Mean-variance between two studies (CCLE-GDSC)

Mean comparison between two studies

```{r}
mplot <- readRDS(file.path(results_dir, 'CCLE_GDSC','fit_within_study','posterior_drug_fits.RDS')) %>% 
  group_by(drug) %>% 
  filter(n_distinct(experiment)>1) %>% 
  ungroup() %>% 
  filter(drug_type == 'broad') %>% 
  mutate(mean = alpha/(alpha+beta), 
         variance = (alpha*beta)/((alpha+beta)^2*(alpha+beta+1)))  %>% 
  select(drug, experiment, mean, variance) %>% 
  pivot_longer(c('mean','variance')) %>% 
  pivot_wider(names_from='experiment', values_from='value') %>% 
  filter(name=='mean') %>% 
  mutate(label = case_when(drug == '17-AAG' ~ 'A',
                           drug == 'AZD0530' ~ 'B',
                           drug == 'AZD6244' ~ 'C',
                           drug == 'Crizotinib' ~ 'D',
                           drug == 'Erlotinib' ~ 'E',
                           drug == 'lapatinib' ~ 'F',
                           drug == 'Nilotinib' ~ 'G',
                           drug == 'Nutlin-3' ~ 'H',
                           drug == 'paclitaxel' ~ 'I',
                           drug == 'PD-0325901' ~ 'J',
                           drug == 'PD-0332991' ~ 'K',
                           drug == 'PHA-665752' ~ 'L',
                           drug == 'PLX4720' ~ 'M',
                           drug == 'Sorafenib' ~ 'N',
                           drug == 'TAE684' ~ 'O')) %>% 
  ggplot(aes(x=CCLE, y=GDSC)) + 
  geom_hline(aes(yintercept=0)) + 
  geom_vline(aes(xintercept=0)) + 
  geom_point() + 
  ggrepel::geom_text_repel(aes(label=label))  + 
  geom_abline(aes(slope=1,intercept=0),lty=2) + 
  theme_bw() + 
  labs(title = 'Estimated mean') + 
  theme(panel.grid = element_blank(),
        axis.text = element_text(size=14),
        plot.title = element_text(size=16),
        axis.title = element_text(size=14)) + 
  coord_fixed()
```

Variance comparison between two studies

```{r}
vplot <- readRDS(file.path(results_dir, 'CCLE_GDSC','fit_within_study','posterior_drug_fits.RDS')) %>% 
  group_by(drug) %>% 
  filter(n_distinct(experiment)>1) %>% 
  ungroup() %>% 
  filter(drug_type == 'broad') %>% 
  mutate(mean = alpha/(alpha+beta), 
         variance = (alpha*beta)/((alpha+beta)^2*(alpha+beta+1)))  %>% 
  select(drug, experiment, mean, variance) %>% 
  pivot_longer(c('mean','variance')) %>% 
  pivot_wider(names_from='experiment', values_from='value') %>% 
  filter(name=='variance') %>% 
    mutate(label = case_when(drug == '17-AAG' ~ 'A',
                           drug == 'AZD0530' ~ 'B',
                           drug == 'AZD6244' ~ 'C',
                           drug == 'Crizotinib' ~ 'D',
                           drug == 'Erlotinib' ~ 'E',
                           drug == 'lapatinib' ~ 'F',
                           drug == 'Nilotinib' ~ 'G',
                           drug == 'Nutlin-3' ~ 'H',
                           drug == 'paclitaxel' ~ 'I',
                           drug == 'PD-0325901' ~ 'J',
                           drug == 'PD-0332991' ~ 'K',
                           drug == 'PHA-665752' ~ 'L',
                           drug == 'PLX4720' ~ 'M',
                           drug == 'Sorafenib' ~ 'N',
                           drug == 'TAE684' ~ 'O')) %>% 
  ggplot(aes(x=CCLE, y=GDSC)) + 
  geom_hline(aes(yintercept=0)) + 
  geom_vline(aes(xintercept=0)) + 
  geom_point() + 
  ggrepel::geom_text_repel(aes(label=label), size=6)  + 
  geom_abline(aes(slope=1,intercept=0),lty=2) + 
  theme_bw() + 
  labs(title = 'Estimated variance') + 
  theme(panel.grid = element_blank(),
        axis.text = element_text(size=14),
        axis.title = element_text(size=14),
        plot.title = element_text(size=16)) + 
  coord_fixed() 
```

Combine.
```{r}
sfig <- plot_grid(mplot, vplot, labels = c('a)', 'b)'), label_size = 16, nrow = 2)
ggsave(file.path(output_dir, 'mean_variance_CCLE_GDSC.pdf'),width=6,height=12,units='in')
```

# Spearman v Accuracy (CCLE-GDSC-CTRP-FIMM-gCSI)

```{r}
intersection.fitted.list <- readRDS(file.path(results_dir,'CCLE_GDSC1000_CTRPv2_FIMM_gCSI','fit_within_study','intersection.fitted.list.RDS'))
twoway_agreement <- tibble()
for (comp in names(intersection.fitted.list)) {
  tibb <- intersection.fitted.list[[comp]]
  curr_studies <- strsplit(comp,'_')[[1]]
  spearman <- tibb %>% 
    select(drug, cell, realized_value, experiment) %>%
    spread(experiment, realized_value) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(spearman = cor(v1,v2,method='spearman')) %>%
    ungroup()
  drug_types <- tibb %>%
    select(drug, drug_type, experiment, posterior_probability_targeted) %>%
    distinct() %>%
    filter((posterior_probability_targeted>.5 & drug_type=='targeted') | (posterior_probability_targeted<.5 & drug_type=='broad')) %>%
    select(-drug_type) %>%
    spread(experiment, posterior_probability_targeted) %>%
    rename(study1_drugtype = curr_studies[1], study2_drugtype = curr_studies[2]) %>%
    mutate(type = case_when(study1_drugtype*study2_drugtype > (1-study1_drugtype)*(1-study2_drugtype) ~ 'targeted',
                            TRUE ~ 'broad')) %>%
    select(drug, type)
  accuracy <- tibb %>% 
    #filter(drug_type=='targeted') %>%
    select(drug, cell, cell_type, experiment) %>%
    spread(experiment, cell_type) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(accuracy = mean((v1 == 'sensitive' & v2 == 'sensitive') | (v1 == 'resistant' & v2 == 'resistant'))) %>%
    ungroup() %>%
    select(drug, accuracy)
  twoway_agreement <- spearman %>% 
    left_join(accuracy, by='drug') %>% 
    left_join(drug_types, by='drug') %>%
    mutate(comparison = comp) %>% 
    bind_rows(twoway_agreement)
}
twoway_agreement %>%
  mutate(comparison = gsub('_',' & ', comparison)) %>% 
  mutate(color = case_when(type == 'targeted' ~ 'D',
                           TRUE ~ 'A')) %>%
  ggplot(aes(x=spearman, y=accuracy)) + 
  geom_point(shape=21,size=2.5,alpha=.75, fill='grey') +
  theme(axis.text  = element_text(size=14),
        axis.title  = element_text(size=14),
        plot.title  = element_text(size=16)) + 
  scale_y_continuous(limits = c(-.1,1),labels = c(0,.5,1), breaks=c(0,.5,1)) +  
  scale_x_continuous(limits = c(-.1,1), labels = c(0,.5,1), breaks=c(0,.5,1)) +
  labs(x = "Spearman rank correlation",
       title ='Common drugs and cells estimated accuracy',
       y = "Accuracy") + 
  facet_wrap(vars(comparison),ncol=5)
ggsave(file.path(output_dir, 'spearman_accuracy_all_datasets.pdf'),width=14,height=7,units='in')
```

# Proportion of drug types and cell types in common (CCLE-GDSC-CTRP-FIMM-gCSI)

Hex of rounded drug types.

```{r}
intersection.fitted.list <- readRDS(file.path(results_dir,'CCLE_GDSC1000_CTRPv2_FIMM_gCSI','fit_within_study','intersection.fitted.list.RDS'))
drug_types_agreement <- tibble()
for (comp in names(intersection.fitted.list)) {
  curr_studies <- strsplit(comp,'_')[[1]]
  drug_types_agreement <- intersection.fitted.list[[comp]] %>% 
    mutate(posterior_probability_targeted = round(posterior_probability_targeted,digits=2)) %>%
    distinct(drug, experiment, posterior_probability_targeted) %>% 
    spread(experiment, posterior_probability_targeted) %>%
    rename(study1_targeted = curr_studies[1], study2_targeted = curr_studies[2]) %>%
    mutate(comparison = comp) %>% 
    bind_rows(drug_types_agreement)
}
p1 <- drug_types_agreement %>% 
  mutate(study1_label = case_when(study1_targeted >= .5 ~ 'Targeted',
                                  TRUE ~ 'Broad effect'),
         study2_label = case_when(study2_targeted >= .5 ~ 'Targeted',
                                  TRUE ~ 'Broad effect')) %>% 
  ggplot(aes(x=study1_targeted, y=study2_targeted)) + 
  geom_hline(aes(yintercept=.5)) + 
  geom_vline(aes(xintercept=.5)) + 
  geom_hex(bins=16) + 
  scale_fill_continuous(type = "viridis", limits=c(0,80)) +
  theme_bw() + 
  scale_y_continuous(breaks=c(0,.5,1)) + 
  scale_x_continuous(breaks=c(0,.5,1)) + 
  theme(panel.grid=element_blank(),
        plot.title = element_text(size=16),
        axis.title = element_text(size=14),
        axis.text = element_text(size=14),
        legend.text = element_text(size=14),
        legend.title = element_text(size=14)) + 
  labs(x = 'Estimated targetedness in dataset 1', y = 'Estimated targeted in dataset 2',
       title = 'Agreement of estimated targetedness')
```

Bar plot of agreement of drug-type agreement.

```{r}
p2 <- drug_types_agreement %>% 
  mutate(study1_targeted = case_when(study1_targeted >=.5 ~ 'targeted', TRUE ~ 'broad'),
         study2_targeted = case_when(study2_targeted >=.5 ~ 'targeted', TRUE ~ 'broad')) %>%
  group_by(comparison) %>%
    summarise(t00 = sum(study1_targeted=='broad' & study2_targeted=='broad'),
              t11 = sum(study1_targeted=='targeted' & study2_targeted=='targeted'),
              t01 = sum(study1_targeted=='broad' & study2_targeted=='targeted'),
              t10 = sum(study1_targeted=='targeted' & study2_targeted=='broad')) %>%
  ungroup() %>%
  mutate(prop_agree = (t00+t11)/(t00+t11+t10+t01),
         count_agree = t00+t11,
         total = t00+t01+t10+t11) %>% 
  select(comparison, prop_agree, count_agree, total) %>%
  mutate(comparison = gsub('_',' and ',comparison)) %>%
  ggplot(aes(x=reorder(comparison, prop_agree), y=prop_agree)) + 
  geom_hline(aes(yintercept=0)) + 
  geom_hline(aes(yintercept=1)) + 
  geom_col(color='black', fill='grey') + 
  geom_text(aes(label=count_agree, y=prop_agree), hjust=-1, size=5) + 
  geom_text(aes(label=total, y=1), hjust=-1, size=5) + 
  theme(panel.grid = element_blank(),
        axis.text = element_text(size=14),
        axis.title = element_text(size=14),
        plot.title = element_text(size=16),
        legend.position='none') +
  scale_y_continuous(breaks=c(0,.5,1), limits = c(0,1.05)) + 
  labs(x='', y='Proportion', title='Estimated drug type agreement') + 
  coord_flip()
```

Combine.
```{r}
sfig <- plot_grid(p2, plot_grid(NULL,p1, NULL, rel_heights=c(1,5,1), nrow=3), labels = c('a)', 'b)'), label_size = 16, rel_widths = c(1.5,1))
ggsave(file.path(output_dir, 'prop_agree_all_datasets.pdf'),width=14,height=6,units='in')
```

# Spearman v Accuracy (PRISM-CTRP-GDSC - including zeros)

```{r}
intersection.fitted.list <- readRDS(file.path(results_dir, 'PRISM_GDSC_CTRP','zeros_included','intersection.fitted.list.RDS')) 
twoway_agreement <- tibble()
for (comp in names(intersection.fitted.list)) {
  tibb <- intersection.fitted.list[[comp]]
  curr_studies <- strsplit(comp,'_')[[1]]
  spearman <- tibb %>% 
    filter(drug_type=='broad') %>%
    select(drug, cell, realized_value, experiment) %>%
    spread(experiment, realized_value) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(spearman = cor(v1,v2,method='spearman')) %>%
    ungroup()
  drug_types <- tibb %>%
    select(drug, drug_type, experiment, posterior_probability_targeted) %>%
    distinct() %>%
    filter((posterior_probability_targeted>.5 & drug_type=='targeted') | (posterior_probability_targeted<.5 & drug_type=='broad')) %>%
    select(-drug_type) %>%
    spread(experiment, posterior_probability_targeted) %>%
    rename(study1_drugtype = curr_studies[1], study2_drugtype = curr_studies[2]) %>%
    mutate(type = case_when(study1_drugtype*study2_drugtype > (1-study1_drugtype)*(1-study2_drugtype) ~ 'targeted',
                            TRUE ~ 'broad')) %>%
    select(drug, type)
  accuracy <- tibb %>% 
    filter(drug_type=='targeted') %>%
    select(drug, cell, cell_type, experiment) %>%
    spread(experiment, cell_type) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(accuracy = mean((v1 == 'sensitive' & v2 == 'sensitive') | (v1 == 'resistant' & v2 == 'resistant'))) %>%
    ungroup() %>%
    select(drug, accuracy)
  twoway_agreement <- spearman %>% 
    left_join(accuracy, by='drug') %>% 
    left_join(drug_types, by='drug') %>%
    mutate(comparison = comp) %>% 
    bind_rows(twoway_agreement)
}
twoway_agreement %>%
  filter(type == 'targeted') %>% 
  filter(comparison != 'CTRP_GDSC') %>% 
  mutate(comparison = gsub('_',' & ', comparison)) %>% 
  mutate(color = case_when(type == 'targeted' ~ 'D',
                           TRUE ~ 'A')) %>%
  ggplot(aes(x=spearman, y=accuracy)) + 
  geom_point(shape=21,size=2.5,alpha=.75, fill='grey') +
  theme(axis.text  = element_text(size=14),
        axis.title  = element_text(size=14),
        plot.title  = element_text(size=16)) + 
  scale_y_continuous(limits = c(-.1,1),labels = c(0,.5,1), breaks=c(0,.5,1)) +  
  scale_x_continuous(limits = c(-.1,1), labels = c(0,.5,1), breaks=c(0,.5,1)) +
  labs(x = "Spearman rank correlation",
       title = 'Common drugs and cells estimated accuracy - PRISM including zeros',
       y = "Accuracy") + 
  facet_wrap(vars(comparison),ncol=5) + 
  theme_bw()
ggsave(file.path(output_dir, 'spearman_accuracy_prism_datasets_zeros.pdf'),width=12,height=6,units='in')
```

# Spearman v Accuracy (PRISM-CTRP-GDSC - excluding zeros)

```{r}
intersection.fitted.list <- readRDS(file.path(results_dir, 'PRISM_GDSC_CTRP','zeros_excluded','intersection.fitted.list.RDS')) 
twoway_agreement <- tibble()
for (comp in names(intersection.fitted.list)) {
  tibb <- intersection.fitted.list[[comp]]
  curr_studies <- strsplit(comp,'_')[[1]]
  spearman <- tibb %>% 
    filter(drug_type=='broad') %>%
    select(drug, cell, realized_value, experiment) %>%
    spread(experiment, realized_value) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(spearman = cor(v1,v2,method='spearman')) %>%
    ungroup()
  drug_types <- tibb %>%
    select(drug, drug_type, experiment, posterior_probability_targeted) %>%
    distinct() %>%
    filter((posterior_probability_targeted>.5 & drug_type=='targeted') | (posterior_probability_targeted<.5 & drug_type=='broad')) %>%
    select(-drug_type) %>%
    spread(experiment, posterior_probability_targeted) %>%
    rename(study1_drugtype = curr_studies[1], study2_drugtype = curr_studies[2]) %>%
    mutate(type = case_when(study1_drugtype*study2_drugtype > (1-study1_drugtype)*(1-study2_drugtype) ~ 'targeted',
                            TRUE ~ 'broad')) %>%
    select(drug, type)
  accuracy <- tibb %>% 
    filter(drug_type=='targeted') %>%
    select(drug, cell, cell_type, experiment) %>%
    spread(experiment, cell_type) %>% 
    na.omit() %>%
    rename(v1 = curr_studies[1], v2 = curr_studies[2]) %>%
    group_by(drug) %>%
    summarise(accuracy = mean((v1 == 'sensitive' & v2 == 'sensitive') | (v1 == 'resistant' & v2 == 'resistant'))) %>%
    ungroup() %>%
    select(drug, accuracy)
  twoway_agreement <- spearman %>% 
    left_join(accuracy, by='drug') %>% 
    left_join(drug_types, by='drug') %>%
    mutate(comparison = comp) %>% 
    bind_rows(twoway_agreement)
}
twoway_agreement %>%
  filter(type == 'targeted') %>% 
  filter(comparison != 'CTRP_GDSC') %>% 
  mutate(comparison = gsub('_',' & ', comparison)) %>% 
  mutate(color = case_when(type == 'targeted' ~ 'D',
                           TRUE ~ 'A')) %>%
  ggplot(aes(x=spearman, y=accuracy)) + 
  geom_point(shape=21,size=2.5,alpha=.75, fill='grey') +
  theme(axis.text  = element_text(size=14),
        axis.title  = element_text(size=14),
        plot.title  = element_text(size=16)) + 
  scale_y_continuous(limits = c(-.1,1),labels = c(0,.5,1), breaks=c(0,.5,1)) +  
  scale_x_continuous(limits = c(-.1,1), labels = c(0,.5,1), breaks=c(0,.5,1)) +
  labs(x = "Spearman rank correlation",
       title = 'Common drugs and cells estimated accuracy - PRISM excluding zeros',
       y = "Accuracy") + 
  facet_wrap(vars(comparison),ncol=5) + 
  theme_bw()
ggsave(file.path(output_dir, 'spearman_accuracy_prism_datasets_nozeros.pdf'),width=12,height=6,units='in')
```
